start rounds
This adds the functionality to start a new round. It will increment the
current round number within the game struct, shuffle and attach a new
deck to the draw pile, and then give each player a hand with 10 cards.

Additionally, I've included a .iex.exs file so that all of the current
modules will automatically be aliased when you run `iex -S mix` and you
will already have a list of players and a game generated for you to work
off of.