create the create/join game screen
This adds the screen that will let a user either create or join a game.
It required updating to the latest version of Phoenix LiveView. While I
was at it, I updated other dependencies as well. There will probably
still be a good deal of refactoring, but this gets that screen created.