cleanup discarding functionality
This makes it so that the app will properly handle an empty discard pile
and differentiates between it and a card being there. It will also
update the discard pile when drawn from and not just when discarded. And
now when someone discards, it will broadcast that the turn now belongs
to the next player