fix card tapping on iOS
For some reason iOS Safari doesn't support the phx-click on div elements
by default, and only seems to work on a and button elements. If you add
an onclick attribute to divs and other elements, it seems to work
though.

This adds an empty onclick method to all of the divs with phx-click
actions to hack our way around the limitation.