sort hands
Sort each player's hand when they first connect to the LiveView, but
again whenever they discard.

This makes it so that the card they draw will always be the first card
in their hand so that it stands out during that turn, but then when they
discard it will get sorted in with the rest of their hand.