update LiveView and add LiveDashboard
This updates LiveView from 0.10.0 to 0.12.1. As far as I can tell, none
of the breaking changes apply to this codebase.

It also adds LiveDashboard, which will allow us to see server metrics
and other things from the /admin/dashboard endpoint using basic auth.
We'll probably want to remove this once we're ready to launch.