allow cards to be played to the table
This allows players to take cards from their hand and play them to the
table. It verifies that the cards are valid for the group they're being
played to and displays them in front of the player. It doesn't currently
save those to the game or display them to other players.