display the scoring page
At the end of a round, all players will now click a button to be
directed to the scoring page where they will be able to see the list of
players and their scores and current levels.

Additionally the page contains a button to mark a player as ready for
the next round, but it doesn't do anything yet.

Resolves #40