show correct round winner
Whenever a player won a round by discarding, the current player would be
changed before the round finished event broadcasted. Because of that the
player whose turn would have been next was shown as the winner, rather
than the player who actually won.

This fixes that bug so that we're showing the proper winner for each
round.