display the number of cards in each player's hand
This adds an indicator to the player list on the game screen so that you
can see the number of cards in each player's hand. This allows you to
see whether or not the player whose turn it is has already drawn. It
also allows the user to know whether or not the round is close to being
over.

Resolves #36