remove outdated test
Previously Game.complete_round/1 was clearing out the artifacts of that
round, but 1161d71cad43aa0cbcd1770337b0cc762b8c4b39 changed that to
allow for players to still be able to see the table until they click the
button to see the scores.

This removes a test that was still expecting the older behavior so that
the CI pipeline will now pass.