use SVGs for rendering cards
The cards were getting rendered at all kinds of different aspect ratios
since there was nothing enforcing that. This uses SVGs to render them so
that they will maintain a strict aspect ratio.

In addition, this adds a new card back for the discard pile that seems
to be more obvious than what was there before.