create end game screen
This modifies the screen for displaying scores to display a proper end
game screen that shows the winning player at the top with all of the
players and their scores beneath. When a player presses the "End Game"
button, they will now be redirected back to the homepage rather than
into a new round.

Resolves #44