store lobby state in the URL
This will update the URL of the page whenever anything is done in the
lobby so that it can be restored if the player refreshes the page or
anything like that.

This will also allow for a URL to be sent to an invitee to allow them to
join the game by only providing their name. Resolves #43