fix card display in Safari
Safari didn't handle the CSS grid properly so card heights were messed
up. This moves to use flexbox rather than grid for displaying the table
so that cards are properly visible in Safari and other iOS derivative
browsers.

It also adds a couple of convenience classes to the IEx helpers so that
you can easily open non-default browsers. You can use `Seeds.safari` to
open the two browser tabs for Safari and `Seeds.chrome` to open the two
browser tabs in Chrome.

Resolves #53