make all wilds and skips black
In more modern versions of comparable card games that players are used
to, wilds and skips have all been made black. This also helps to
emphasize that these are special cards.

This changes all of the wild and skip cards in Level 10 to now be black
to match what players will be used to. All examples and tests have also
been updated to reflect this change.