sort cards on the table
Whenever cards are displayed on the table (runs specifically) out of
order, it can be really hard to figure out what cards can be played on
top of them.

This sorts the cards in each group on the table so that it's easier to
tell what cards can be played at any given time.

Resolves #54