use custom Agent
In order to hand off state when a sigterm is received, we need to be
able to have finer control over the GenServer functions so that we can
handle the terminate callback.

This switches out Agent for Level10.Games.GameServer, which is basically
a copy of the implementation of Agent. This should only be a temporary
fix as it makes more sense to use a GenServer for holding the game state
in the long term, but it allows us to get to state handoff much faster.