handoff state at sigterm
This creates a CRDT process on each running node that will connect to
all others. Whenever a SIGTERM is received, the state for each game will
be added to the CRDT.

When the node has been shut down, Horde will attempt to start up the
process again on a node that is still up. At that point, the new game
server process will check to see if a game with the given join code
exists in the CRDT. If it does, it will get the game state out of the
CRDT and continue running the game as if nothing had happened.