handle name conflicts
The distributed Supervisor will detect a game server going down whenever
the node is terminated, so it will try to start that game server on a
new node. When it does that, the new game server will pull the game
state out of the StateHandoff CRDT. However, it appears that sometimes
this all happens before the distributed Registry realizes the process is
no longer up, so the registry will kill the new game server because of a
name conflict. When this happens currently, it kills the process without
sending the state into the CRDT, which means that the game is gone
forever.

With this change, we handle the name conflict exits just like we do
shutdown exits and put the state back into the CRDT so that it can be
pulled out at a later time.

Additionally, we currently put the state into the CRDT for games that
are shut down with Level10.Games.delete_game/1. Games that are deleted
manually shouldn't be restarted, so this commit also prevents those from
being added to the CRDT.