wait after handing off state
State wasn't being passed correctly between nodes. I believe this may be
because the state wasn't properly propagating to other nodes in the
cluster before the application terminated.

This adds a short sleep time once state has been added to the CRDT so
that it has time to propagate to the rest of the cluster.