only add to the crdt if the game doesn't exist
Whenever we're running state handoff, it appears that sometimes a new
game gets created before the old one has synced into the local CRDT.
Usually that node will then be terminated and put an empty game into the
CRDT on top of the good one that was already there.

Really we ought to ensure that a CRDT has synced before allowing the
process to die, but this is an attempt at a workaround by making sure
that a game doesn't exist in the CRDT yet before adding it.