monitor nodes in the cluster for state handoff
Previously, we were only updating the CRDT neighbours when a new node
joined the cluster. When a node left the cluster, the other CRDTs
wouldn't notice that it left and would still try to sync with it.

This uses :net_kernel to monitor nodes joining and leaving the cluster
so that we can add and remove neighbours from the StateHandoff CRDT in a
more robust manner.