store levels separate from scores
We uncovered a bug where players who weren't connected to the game
screen when a round was completed could see an "Internal Server Error".
Upon investigation, it appeared that this was because players were
completing the level they had been working on that round. When the round
was completed, scores would be calculated and levels would be
incremented. When a player missed the message to continue on to the
scores screen, they could potentially load the game screen with the new
level information and it wouldn't match the table information for that
round.

This stores the levels separately from the scores (while keeping them in
the scores too). The levels in the scores will be incremented when the
round is complete, but the separate levels will only be updated when the
new round starts. This way the state will remain in sync and there
shouldn't be any issues with it.