fix bug with empty level requirements
Whenever a new game was started, its level requirements were empty. This
is because we were updating them using the game's scores, but for new
games the scores weren't in place yet at that point.

This puts the empty scores on a new game before updating the levels so
that the levels can retrieve the information they need.