switch up player going first
Every round we reset the turn counter to 0. Because this was the only
thing that was determining player turn, the same player would go first
every round.

The algorithm has been updated to now take the round number into account
as well so that a different player will go first each round.

Resolves #57