redirect scores to game when round started
Previously, if a player wasn't on the scores screen when the round
started, they would miss the PubSub message and be stuck on the scores
screen unless they manually edited their URL.

This checks if the round has started, and if so, redirects any player
who goes to the scores screen over to the game screen.

It also starts to move toward a pattern of only getting the GenServer
state once and getting everything necessary out of it rather than
sending a message back and forth for each field that is needed.

Resolves #61