fix vertical height and padding
On iOS Safari, users were having to scroll the screen to get to the
buttons. This is because iOS Safari made a conscious decision to measure
the vertical height of the screen based on the maximum height. This
maximum is only reached after a user has scrolled and Safari's UI
disappears.

This shouldn't have to happen, so this commit measures the inner height
of the screen and uses that to calculate how tall to make everything. It
also adjusts some of the other elements to make them look good within
the new shorter screen.