sort wilds appropriately
This creates a new function: Levels.sort_for_level/2 as well as a lower
level counterpart: Card.sort_for_group/2 that will sort the cards in a
level group with the wilds placed in the correct spots.

Additionally this mostly solves #59 by not allowing wilds in a run to be
replaced. However, this only blocks replacement of wilds in the middle
of a run. Wilds with no place inside the run can still be used on either
end of it, and won't be locked in until another card is played that
requires the wild to act as a certain number.