configure for kubernetes
This does the work of moving from Gigalixir to Kubernetes. All of the
k8s configuration is located in the k8s folder. This also updates the
libcluster strategy to use Kubernetes' API for clustering, and adds a
health check endpoint for readiness checking.

We will want to later update the Github Actions workflow to deploy to
k8s as well.

Resolves #70