improve end game
The end game was a bit broken. This was because the live view was
looking for the game's status to be set to :finished, when it actually
gets set to :finish. This commit fixes the live view so that it will now
expect the correct value.

In addition it swaps out the (11) for a trophy emoji whenever a player
has completed all of the levels since the 11 didn't really make sense
there.