update skipping mechanism
Previously whenever a player discarded a card, the app would check to
see if that card was a skip. If it was, the app would then skip over the
next player.

We want to allow players to choose which player in the game should have
their next turn skipped. This helps to even out the playing field a bit
and create closer games.

This doesn't yet give them the choice, but it moves in that direction.
We create a new set on the Game struct that holds the IDs of players who
should be skipped on their next turn. Then, rather than skip them
immediately, the chosen player ID can be added to this set. Whenever a
player's turn is over, the application will check if the next player is
in this set of skipped players before letting them take their turn.

This is fully backwards compatible by adding the ID of the next player
into the set. First half of #74