fix tailwind purging in docker
The CSS styles weren't being properly processed in production, but
everything was working fine on my local machine. I realized this was
because the lib folder wasn't present for Tailwind to use when purging
unused styles. This commit adds the lib folder to the Node docker image
that's used for building the CSS file so that all the classes will be
present and used for purging.