use database users for starting games
This is the big one. It will use the currently logged in user to create
a new Player struct that can be used for creating or joining a game. And
it removes the player IDs from the URL so that you can't pretend to be
someone else just by getting their URL. This makes Level 10 a lot more
secure.