authorize the live dashboard only for admins
Currently a user can use the live dashboard to see the log lines that
result from any actions they take. This hasn't been an issue in the past
because there was no secure information going through the server, but
now that users can log in, this needs to be locked down.

This adds a user_role enum to the database and a role to all the users
in the database. Then it only allows users with the admin role to access
the live dashboard.

Fixes #81