use execute/2 for creating types
This modifies the migration that created user roles. The type was
originally written using execute/1 which required separate up and down
functions. This changes it to use execute/2 instead to be able to use a
single change function. This creates a little more safety for properly
running rollbacks. Normally I wouldn't change a migration that has
already gone through, but in this case I wanted to use the proper
pattern so that it can be referenced in the future.

Additionally this aliases the accounts modules for IEx so that they can
be used more easily.