add k8s setup file
When setting up a new Kubernetes cluster to run Level 10, some extra
pieces are needed. Namely a service account with the proper roles bound
and the necessary secrets. This adds a new setup.yaml file that can be
run to take care of all of that. It also adds the various Kubernetes
tools I use to the .tool-versions file so that asdf will use the same
versions if anyone else is running it.