add local deployment script
This adds a new script in the bin folder that can be used for easily
deploying to a (probably) local Kubernetes cluster. It will build a new
Docker image, tag it with the current UNIX timestamp, then tell
Kubernetes to set the Level 10 image to the newly built one. Finally it
will watch the rollout status to make sure it deploys as expected.

This should make it much simpler to test deployment locally.