use pods for clustering
Currently k8s endpoints are used for clustering. This means that a node
will be removed from the cluster as soon as Kubernetes removes it from
the list of endpoints. I think this might be the reason state sometimes
doesn't get added to the handoff CRDT before the node is unavailable.

This is potentially the answer to #67