make the auth emails pretty
This uses Bamboo.Phoenix and Maizzle to create new HTML email templates
for the auth emails. Using Maizzle allows the use of TailwindCSS for
styling so that the emails match the theme of the app pretty well. I'm
unsure of how to actually include the Maizzle stuff in this repo though,
so it lives separately for now.

This resolves #82