get only the fields the player needs
While using Games.get/1 to fetch the game simplifies the code greatly,
it's also sending a lot more info over the wire than what the player
actually needs. This adds a new Games.get_for_player/2 function that
will get rid of the fields the player doesn't actually need and send
less data over the wire. In limited testing, it appears that the full
game is a little over 10kB and with the unnecessary fields removed it's
about 1.6kB so about a 6x improvement.