refactor current_state out of after_join
This creates a new current_state function separate from the after_join
function. This will allow for sending the full state at any time rather
than just when the socket is mounted.

This commit also contains a lot of other changes. Most of the
dependencies are updated, the Elixir and Erlang versions are updated,
and dialyzer errors are fixed throughout the application.